requirements.txt
apache-airflow~=3.0.2
pyarrow>=15,<18
boto3==1.40.1
botocore==1.40.18
fastapi~=0.115.14
pandas~=2.3.0
pendulum==3.1.0
Requests~=2.32.4
#Database
SQLAlchemy~=1.4.54
psycopg2-binary==2.9.9
apache-airflow-providers-postgres
#
redis~=5.2.1
s3fs~=2025.9.0
eurostat==1.1.1


Jinja2~=3.1.6
uvicorn~=0.35.0
psycopg~=3.2.10
seaborn~=0.13.2
style~=1.1.0
pip~=25.2
attrs~=25.3.0
protobuf~=5.29.5
cryptography~=45.0.4
pillow~=11.2.1
docutils~=0.16
Sphinx~=5.3.0
futures~=3.0.5
tornado~=6.5.1